Move number of islands to image.morphology, add function to calculate morphology featueres
