+++ ctapipe/image/__init__.py
@@ -1,12 +1,19 @@
 from .hillas import (
     hillas_parameters, HillasParameterizationError, camera_to_shower_coordinates,
 )
+from .timing import timing_parameters
+from .leakage import leakage
+from .concentration import concentration
+from .statistics import descriptive_statistics
+from .morphology import (
+    number_of_islands,
+    number_of_island_sizes,
+    morphology_parameters,
+)
+
 from .cleaning import *
 from .pixel_likelihood import *
 from .extractor import *
 from .reducer import *
-from .muon import *
 from .geometry_converter import *
-from .leakage import *
-from .concentration import concentration
-from .statistics import descriptive_statistics
+from .muon import *
