+++ ctapipe/image/timing.py
